Bugfix: new oauth redirector was impossible to use correctly.
Problem:
In a previous change, I broke out some common code into a Redirector
object, that just checks a cookie and sends a redirect in case of error.

I did not notice that the code to check for redirects was invoked before
validating the cookie, thus causing the code to always faill (eg, after
a succesful authentication, the code would return an error as the url
redirected back to the authentication destination).

In this change:
- fix the code so that it's actually usable.