Refactoring of initialization code.
Background:
CLI commands are made by a main command and subcommands.
Most commands and subcommands share some options.

For example: all the astore commands need to define an astore
server. Most commands require the authentication token for the
user, or a logging object.

Before this change:
Initialization of some of those common objects was disjoint.
So for example, the base library was initializing a logger,
as was astore. With different CLI flags, and great confusion
when those things were shared.

After this change:
A lot of the initialization logic has been moved in common blocks,
shuffling the code around significantly.
There are still some dark corners, but it's a step in the right
direction.

This means that:
- one logger is now initialized for everyone.
- flag reloading has a consistent behavior.
- one object to manage identity, used by everyone.

A few usability bugs addressed:
- correct help message now printed in case of error.
- login error now handled consistently, provides guidance to the user.
- logging now happens consistently.

In terms of dark corners / todos:
- wrong flags to bootstrap libraries (eg, logging) can end up ignored.
  This is fixable, just not in this change.
- identity access libraries don't cache the identity, reload the
  file at each invocation.

Additionally, run go fmt through the code base.