Use errors.As rather than direct type conversion in retry library.
Before this change:
retry functions were using err.(*whatever) to check if an error
was of a specific type (eg, FatalError).

Problem:
With go 1.13 and the new interfaces to wrap and unwrap errors,
that way of checking error types is deprecated. Further, it
prevents the library from detecting the correct error type under
a number of conditions.

With this change:
Use the new errors.As facilities to do error conversion.