astore: save metadata on gcs.
Before this change:
Uploaded files were just blob of data, no metadata associated apart
from the default GCS metadata.

After this change:
Whena a blob of data is committed as a file, we save the uid and path
of the file pointing to that blob. This is best effort, but can
help recover data, and clean up the bucket.