testfix: remove flakiness of remote_test
remote_test.go is checking that a returned array has the correct
content. The order of the elements is irrelevant, but the test
was comparing a fixed array. The content of the array was from
a map, order is unpredictable.

In this change, we just verify that the set of elements is correct,
regardless of the order.