nassh: verify credentials token even at connect time.
Background:
- when connecting for the first time, a nasshp proxy validates your
  authentication cookie, and returns in exchange a signed sid.
- the signed sid can be used at any time to establish and re-establish
  a connection to a given backend.

Problem:
If the user no longer has an authentication cookie, account terminated, ...
The SID is still accepted. Further, it is accepted for as long as the
server uses the same encryption key.

This was a known issue, and this change finally addresses the TODO.
It's a simple change, it just adds a few lines to verify the cookie
at connect time.

Next: expire websocket connections based on validity of authentication cookies.