nassh: allow access to proxy that have authentication disabled.
Before this change
Tunnel libraries would refuse to even try to connect to a remote
proxy without an authentication cookie.

The main reason for this behavior was that until a few changes ago,
it was not possible to tell authentication errors apart from other
errors, causing the client to keep retrying with unclear error messages.
However, previous changes added the capability to detect authentication
errors, and the client as is today handles them just fine.

In this change
Tunnel libraries allow to connect to the proxy even without an
authentication cookie.