enproxy: allow ssh gateway and oauth proxy to share the same host.
Before this change
The ssh gateway and http proxy had to be assigned to different virtual
domains / different hosts.

With this change
The two can share the same mux, although in a hacky way. This allows
to keep the two either on different domains (exactly like before) or
on the same domain.