bugfix: do send acks if there is no write for x amount of time.
Background
The browser writer function is supposed to write to the browser when
a) there is new data to write (doh) and b) when there is read data
to acknowledge. Similar to nagle, the current code does not acknowledge
data immediately. It acknowledges on first write, or after a timer expires
waiting for data to be written.

Problem
WaitFor() waiting for data or a timer to expire was buggy. When the
timer expired, it returned nil rather than indication of expiration.
The caller was thus never seeing the timer event, and never sending
the acks outside of writes.

In this change
Fix WaitFor to return the proper error.