nasshp tunnel: implement better timeouts.
Before this change, we had the skeleton to configure timeout, but
only a handful were actually implemented in the tunnel client.

This change finally implements more granular timeouts, alongside
the ping logic. Specifically:
1 - implement connection write timeout, when the underlying connection
    supports it.
2 - send a periodic ping to ensure the remote server is reachable.
    If no response is received, time out the reads.

This second improvement is fairly important. Without it, it would take
10s of minutes for the client to realize the connection was dead.

The reason is that the original code only had a write deadline, but
writes succeed for as long as there is space in the tcp buffers, so
in practice they would never fail.

Setting a static timeout was also bad, as it's impossible to predict
when the new read would have data, even on a healthy connection.
The new code is very simple, and is entirely based on sending a periodic
ping, and waiting for a reply.

Additionally:
- implement better error handling. If we encounter a fatal error
  with the server, there is no point in retrying.