Build: drop docker rules, no longer required.
Additionally: downgrade gazelle.

Reason: the workspace rule to configure gazelle never worked.
Turns out that the docker rules were importing an older version
of gazelle, and that's the version that was being used.
The newer version actually does not work.

Downgrading was the only way to keep this change a noop.
Next changes will update the version used.