khttp: add function to read responses in a streaming fashion.
Background:
The existing API allows to easily collect the result in a variable, or
write it in an object. But every streaming write, requires additional
complexity and an extra copy.

This change introduces an alternative reader that allows direct streaming.
The old API could probably be absorbed by this new one, but will leave
this clean up for later.