kflags/env augmenter: make the prefix optional.
Background:
When setting flags from the environment, the first prefix automatically
added by the client library corresponds to ARGV[0] of the command, the name
of the command, so eg, a flag "auth-server" is looked up as ENKIT_AUTH_SERVER.

Before this change:
The env augmenter required a prefix to always be passed. So for example if
the caller supplied "enkit" as prefix, the variable ENKIT_ENKIT_AUTH_SERVER
would be looked up.

In this change:
Make the prefix entirely optional, it is up to the caller to provide.