lib/cache: provide a method to access existing keys only, without allocating one.
Background:
Right now, the cache API only allows to atomically get or create a key, and
separately to check the existance of a key, without allowing access.

A bug fix in kflags/kconfig has a first legitimate case where if a key exists
already, we want to access its value. If it does not exist, however, we need
to do something much more complex, that requires using the atomic Get operation.

In this change:
- modified the existing Exists API to allow for read only access to an existing
  key. This significantly limits the implementation options, as now there's no
  longer a Commit/Rollback at the end of the work. But so is life, we'll fix this
  when / if we have a second implementation of a dir cache.