bugfix: allow logging to keep working past initial setup.
Background:
Until flags are parsed, enkit does not know how to log data.
Still, it has to log data. So, it uses an accumulator / temporary
logger, to then start using the real one.

Problem:
Some messages never appeared in the log. Turns out that objects instantiated
during flag parsing time may take a reference to the temporary logger, and
use it indefinetly. Causing messages to get discarded, and well, a memory leak.

In this change:
Introudce a "Proxy" logger and start using it. A proxy logger internally just
forwards any log request to a different logger.  When flags are parsed, rather
than replace the proxy logger itself, replace the internal logger.

This means that any object with a referecne to the proxy logger will start
seamleslly using the new logger.