bazel/linux: Specify module name in make command
Specify the module name to the make command used internally by the
kernel_module rule (+ refactor the command to make it easier on the eyes).

Our current make command works fine in cases where we define a single
kernel module in the Makefile/Kconfig file. Instead, in cases where
there are multiple modules defined in a single file (i.e., multiple
objects assigned to obj-m), it ends up compiling all of the modules
instead of the one specified my the kernel_module rule.
Specifying the module name, forces the Kbuild system to only build that
module.