bazel/linux: Add KernelModuleInfo provider
Introduce the new KernelModuleInfo provider. It is exposed by
kernel_module_rule and aims at providing rules depending on the kernel
module info about how the module has been built.
Currently it has no uses, but it will be useful for the new rule
handling kernel testing. Foreseeing that, it exposes the generated .ko
file and the kernel tree the module has been compiled against, so that the
kernel testing environment can confirm compatibility with the linux
executable image that will be used.