bazel/linux: Add the new kernel_test rule
Add a new kernel_test rule which replace the previosly used macro with
the same name.
This new rule leverages the info contained in the new providers, exposed
by the new repository rules used to retrieve the executable kernel image
and rootfs image (RootfsImageInfo and KernelImageInfo), to retrieve the
files needed to run a kernel test and to make sure that all the pieces
provided are compatible with each other.
As for the script handling the test execution, move to a template based
generation, which is more reliable and less bug prone.