KUnit-bazel integration: use providers and rules instead of a macro wrapper (#20)
* bazel/linux: Fix typos

* bazel/linux: Add KernelModuleInfo provider

Introduce the new KernelModuleInfo provider. It is exposed by
kernel_module_rule and aims at providing rules depending on the kernel
module info about how the module has been built.
Currently it has no uses, but it will be useful for the new rule
handling kernel testing. Foreseeing that, it exposes the generated .ko
file and the kernel tree the module has been compiled against, so that the
kernel testing environment can confirm compatibility with the linux
executable image that will be used.

* bazel/linux: Remove deprecated kernel_test macro

Remove the deprecated macro that will be substituted by a bazel rule in
the following patches.

* bazel/linux: Add the kernel_image_version rule

This repository rule mimics the already available kernel_tree_version
(previously named kernel_version), but provides a precompiled linux
kernel executable image instead of a kernel tree to compiled out of tree
module against.

kernel_image_version, internally generates a kernel_image target with
the same name, which exposes the new KernelImageInfo provider. This new
provider provides the necessary info which will be consumed by the
kernel_test rule (added in the following patches).

* bazel/linux: Add the rootfs_version rule

The approach is the same used for the kernel_image_version rule. See the
previous commit for a quick overview.

* bazel/linux: Add the new kernel_test rule

Add a new kernel_test rule which replace the previosly used macro with
the same name.
This new rule leverages the info contained in the new providers, exposed
by the new repository rules used to retrieve the executable kernel image
and rootfs image (RootfsImageInfo and KernelImageInfo), to retrieve the
files needed to run a kernel test and to make sure that all the pieces
provided are compatible with each other.
As for the script handling the test execution, move to a template based
generation, which is more reliable and less bug prone.

* bazel/linux: rename kernel_version to kernel_tree_version

For naming consistency reason with the newly introduced
kernel_image_version repository rule (representing an executable linux
kernel image), rename the kernel_version (representing the source code
to build out of tree kernel modules) to kernel_tree_version.

* bazel: Apply //:buildifier linting

* bazel/linux/defs.bzl: Fix typo

* bazel/linux/defs.bzl: Fix typo