bazel/linux: Better handling of kernel_test error
The kernel_test rule fails if the test kernel module and the linux
kernel executable image provided are not compatible (it checks that both
have been generated by the same kernel_tree package).
In case of failure, instead of returning an empty DefaultInfo provider,
use the better suited fail function.