kbuild/utils/README: Add hint to kernel_test rule
In the workflow section, add an hint to the fact that during the specified process we have generated a linux kernel executable image as well. Since most generate_custom_archive.sh users will be kernel_test rule users, it makes sense to point that out and link to the kernel_test rule documentation.