bazel/utils/transform: add support for "tree artifacts".
Background:
Bazel rules can generate "file artifacts", or "tree artifacts".
A file artifact is 1 bazel object, representing a file.
A tree artifact is 1 bazel object, representing a tree of files.

Individual files in tree artifacts are invisible to the rest of
the bazel rules, as they don't have a corresponding file object.

In this change:
- modify the transform rule so that it can handle "tree artifacts",
  and transform files from tree artifacts as per "include" and
  "transform" parameters.

- add support for the "expand" attribute, so "transform" can be
  used to break a "tree artifact" into independent "file artifacts"
  or sub-"tree artifacts".

Additionally:
- Improve labels.bzl to handle paths that start fromt the root
  of the label (wrong result was returned before).
- Add an helper function in match, to turn a simplified pattern
  into a valid shell pattern.
- Add a debug parameter to transform, so one can see what the
  rule is doing exactly.

Getting this to work was a bit tricky:
- as the content of tree artifacts is inaccessible to bazel,
  the rule now generates a shell script to transform and
  include the files from tree artifacts.