bazel/linux/defs.bzl: Fix kernel_module bug
As a default "srcs" parameter, consider only the makefile parameter and
all the .c and .h files belonging to the package where the kernel_module
rule has been instatiated.

Currently, if the user doesn't specify the "srcs" parameter of the
kernel_module rule, the implementation will consider all files belonging
to the package to be the source files for that kernel module.
If for some reason, a file with the same name of one of the intermediate
build files pollutes the package, bazel will try to copy/link that over in
the build dir. This seems to cause a compilation failure since that file
could already exist there and that filesystem is read-only?

The bug can be reproduced, for example, by copying the module_name.mod.c
in the package source tree and modifying it (or the kernel module sources).
The error will look like this:

redacted: Read-only file system
make[2]: *** [redacted/Makefile.modpost:111: redacted/Module.symvers] Error 1
make[1]: *** [redacted/Makefile:1772: single_modpost] Error 2
make: *** [redacted/Makefile:185: __sub-make] Error 2