[bugfix]: pam_script_acct should exit 0 if autouser is not involved.
Before this change: if AUTOUSER NSS was not involved with the login
of a user, the pam script would fail. This would cause existing users
in the normal passwd file to not be able to log in.

After this change: if AUTOUSER NSS was not involved with the login,
exit 0, to let pam continue with whatever it is doing, rather than
kicking the user out.