kbuild: update custom archive script for use with enf kernel builds
The Enfabrica kernel build separates the kernel config from the kernel
build output directory.  Previously, the archive script assumed the
kernel config had a specific path.  Thus the archive script and
Enfabrica kernel build layout were incompatible.

The change here is to add an option to the archive script to specify
the path of the kernel config.  The default maintains the previous
behaviour, assuming ".config" in the kernel build directory.

Also the kernel version parameter is now optional.  If not specified
the kernel version is taken from the kernel build directory, from
${SRCDIR}/include/config/kernel.release.

Signed-off-by: Curt Brune <curt@enfabrica.net>