Add NVidia driver build rule
Add a build rule for building the NVidia driver family of kernel
modules.  The NVidia kernel module build does not follow the model of
a stock kernel module build, so create a custom rule for it.

Along the way refactor the regular kernel module build rule to support
the original use case and the NVidia case by adding support for:

- building multiple kernel modules

- providing custom arguments to the Makefile

Signed-off-by: Curt Brune <curt@enfabrica.net>