lib/oauth: add tiny cli tool to generate and print tokens.
This PR creates a tiny utility to manually inspect and generate tokens.

Together with this utility, there is a minor refactor of lib/oauth/factory.go
to allow instantiating the code to actually sign and generate cookies.

The utility takes a set of keys from command line, and can either
open cookies, or create them, with arbitrary duration.