kflags/kconfig: export enkit_path to running packages.
Background:
When a package is downloaded from our internal repository and run,
a few variables are passed via the environment and can also be use
for command line flag substitution.

In this PR:
Add the 'enkit_path' variable, replaced as '{enkit_path}' on the
cli, or exported as KCONFIG_ENKIT_PATH on the environment.