feature/fix: disable TLS by default with backends. (#111)
This is a temporary hack waiting the overhaul of the proxy
configuration file. Until then, let's disable certificate
verification when talking to internal backends.

This is a security risk. However, most of the interal backends
configured happen to be horrible ssl appliances with self signed
certificates, and no ability to configure a real certificate.
Practically speaking, this is the only option to get them to
work.

With the config overhaul, it will be possible to selectively
disable certificate validation, or pin certificates directly
in the configuration file.