lib/server: break out assets management function. (#133)
This is a minor cleanup: right now, server.go includes a Run()
with heavy dependencies on grpc, and a bunch of helper functions
to deal with asset files.

The functions for asset file managmenet are generically useful
and self contained. But by being in the server library, they
cannot be used without bringing in the entire grpc ecosystem.

Also, it does not make sense for those functions to be in lib/server.

In this PR:
- move asset management functions as they are in lib/khttp/kassets.
- update code using those functions/libraries as necessary.
- rename lib/server/main.go in lib/server/run.go as there is no main
  there.

This PR is otherwise a noop: no code change has been performed.