proxy/: add ability to remap headers in case sensitive way. (#138)
Long story:
RFCs mandate how capitalization is handled in HTTP headers.
Specifically, mandate that headers have "Name-Whatever-Value"
kind of capitalization. "Name-WhatEver-Value" is not legal,
as "name-whatever-value".

golang base libraries use RFC compliant capitalization in
all outbound requests. So ...-WebSocket- headers are rewritten
as ...-Websocket-...

RFC compliant web servers accept them, but, non-RFC compliant
webservers may choke on them.

In this PR:
- add a configuration parameter to force golang web server to
  be non compliant, by allowing to configure a header to map to
  a differente value in a case sensitive way.

Unfortunately, there is no way to just preserve the case
provided by the client in a sane way, as the golang libraries
sanitize the headers in the first few lines of code of
request handler, and the original case is not preserved.

See equivalent bugs on caddy and traefik web servers.