lib/kflags: add ability to provide command line flags from a map.
This change introduces a new Augmenter, the MapAugmenter, which allows
to provide flag defaults from a simple map, provided at the agumenter
creation time.

The MapAgumenter can be configured to use one or more EnvManglers
(now renamed VarMangler) to allow to compute a different key from
a flag name. For example, by using a Camel mangler, for a flag named
"network-timeout", the key "NetworkTimeout" is looked up in the flag.

Additionally, this PR restructures manglers to be more genereically
useful, and introduces a couple new ones.

No functional change to the codebase, this just adds new library
functions that will be used in an upcoming PR.