lib/khttp/kassets: allow indexes in subdirectories.
Background:
Static assets are served differently depending on the file name.

For example: the .html extension is normally stripped, so foo/bar.html
is served as foo/bar.

Before this change:
/index.html was served as both /index and as /, but a foo/bar/index.html
was only served as foo/bar/index.

After this change:
foo/bar/index.html is also served as foo/bar/. Arguably, this PR fixes
an existing bug in the code, as the intended behavior was for any
index.html file to serve as the directory index.