testing: Add errdiff lib
This change adds a testonly library called "errdiff" that allows unit
tests to easily see if errors match an expected substring.

This allows for a bit more fidelity than checks by the assert package,
as the latter only has functions for checking error presence (though
oftentimes the error expected in tests is not the one actually returned)
and exact error equivalence (though it may be difficult and/or brittle
to formulate an expectation on an exact error value).