git: Add Go lib with temp worktree generation (#235)
This change adds a git library, with initial support for generating
worktrees from a provided repo path.

The bazel changed target detector will need the ability to run at two
different points in the source history locally. One way to do this is to
issue checkout commands in the user's repository; however, it is
possible that a crash/error will leave a user's repository in an
unexpected/unknown state.

Instead, we can shell out to `git worktree` to create essentially a
shallow clone at a particular committish in a different directory (in
this case, a temporary directory) and perform bazel queries on these
temporary worktrees. Here the only risk is that we fail to delete the
temp worktrees upon crash/error, which are trivially cleaned up later
manually or via reboot.

go-git does have the ability to traverse a worktree, but it does not
appear to be trivial to spill a particular commit to disk so that we can
invoke `bazel` on it.

This change adds a small library that shells out to `git worktree` to
add/remove the worktree instead.

Tested: Unit tests in this commit

Jira: INFRA-52