bazel: Parse workspace_log during query
This change modifies Query to return a struct of query results. These
results include targets that match the query, as well as other optional
information. The first useful such piece of information is the workspace
log, which details when:

* WORKSPACE caused a download to happen, including the SHA256 of the
  download
* WORKSPACE caused a command execution to happen

among other events.

These events will be useful in reducing external dependencies down to
the SHA256s of their corresponding downloads, rather than needing to
resort to hashing individual files of the external dependencies.