enkit bazel affected-targets: Add repo_root flag
This change allows the commandline to take a manually-specified
repository root, rather than always using the present working directory.

This will be necessary when running enkit in Cloud Build from a Docker
container, as the Cloud Build file will need to direct enkit to where
the repo is mounted in the container.

Tested:

```
~/dev/enkit/bazel-bin/enkit/enkit_/enkit \
    bazel \
    affected-targets \
    list \
    --start=master~1 \
    --end=master
~/dev/enkit/bazel-bin/enkit/enkit_/enkit \
    bazel \
    affected-targets \
    list \
    --repo_root=/home/bminor/dev/enkit \
    --start=master~1 \
    --end=master
```

Jira: INFRA-52