bazel presubmit: Don't fail when no targets provided
Bazel will exit with an error if `bazel test` is invoked with an empty
target list. This change checks both files for targets, and skips the
respective `bazel build`/`bazel test` step if the file is empty. The
log output is also improved to make it clearer when a step will be
skipped.

Jira: INFRA-113