INFRA-120: Fix affected targets detection with generated files
Prior to this change, affected targets detection would attempt to read
and hash generated files, which don't actually exist.

The purpose of the hash is to detect when targets change. Generated
targets are based solely on the rule that generates them, so this change
ensures that:

* the generating rule is registered as a dependency of the
  generated-file target
* hashing the generated file target doesn't attempt to read any files

Tested: Added unit test for this case

Jira: INFRA-120