INFRA-121: Fix bazel graph parsing for directory labels
Some labels in `internal` are actually directories. When this is the
case, these directory labels are used as commandline flags rather than
file contents. It is debatable if this should just be fixed in
`internal`, but in the meantime dependency detection can tolerate this
case.

When a "source file" label that actually represents a directory is
encountered, the hash of the "source file" should correspond simply to
the hash of the label itself. This resolves a "can't read: is a
directory"-type error when attempting to hash the underlying "file" for
a directory source file target.

Jira: INFRA-121