flextape: Return queue position to client
The client should log to the user the status of the license acquisition
progress when queued. Currently, there's not much progress the client
can communicate, other than that it's queued.

This change modifies the server to return the queue position so that
clients can communicate this to the user, providing user visibility that
in fact progress is being made (or the queue appears to be stalled).

Tested: Unit tests

Jira: INFRA-139