lib/token: add multi encryption support.
Add support to generate tokens that can be decrypted by different
keys.

This is in preparation for future code, but it allows to generate tokens
meant for specific systems to consume them, while issuing a different
key to each such system.

It can also be used to implement schemes where data is stored with
an escrow key alongside a per-user/per-session key, or where data
is encrypted to be sent to different recipients.