lib/config/director: minor bugfix - leaking file descriptor.
This PR fixes a minor file descriptor leak homedir Write, where
a tempfile is created, but the corresponding file object is never
closed.

This is generally not impacting, as configs are read/written to
at startup only.

But the config store API effectively provides a key value store,
and this could be used outside initial configuration handling.