kbuild/v2: split the archive create process into two steps
We need two types of kernel archives.

For bazel builds we need an archive of kernel headers for building
out of tree kernel modules.  Also this archive must be made public to
work with bazel.

For making VMs we need an archive of a Debian repo containing .deb
packages.  This archive should remain private as it contains our
compiled kernel artifacts and kernel source .deb packages.

This commit splits the original workflow into two archives.

Signed-off-by: Curt Brune <curt@enfabrica.net>