lib/khttp/krequest: allow to override the host header.
(this is in preparation for follow up PRs)

Background:
When fetching a resource via HTTP, the client needs to know the
IP of the server, and the "name of the server" that will be used
in the "Host:" header to retrieve the server.

Typically, both the IP and Host header are computed from the URL
to retrieve. Eg, "get http://www.google.com/" will resolve www.google.com
to an IP, and add a Host header containing "www.google.com".

The golang libraries default to the same behavior: compute the
Host and IP from the URL to retrieve.

This is a problem in testing, when connecting to a reverse
proxy, or set of load balanced backends, eg, when the IP resolution
does not match the host that needs to be retrieved.

In this PR:
- add the ability to override the host independent of
  the URL to be retrieved.