New Kernel Artifact Builder (#356)
* kbuild/v2: a new approach to building kernel artifacts

This is the initial commit for a set of scripts to build kernel .deb
packages and upload them to astore.

Signed-off-by: Curt Brune <curt@enfabrica.net>

* kbuilder/v2: add a .gitignore to hide local configuration files

The example build runner allows for including a local configuration
file.  The local configuration file should not be checked in.

Signed-off-by: Curt Brune <curt@enfabrica.net>

* kbuild/v2: split the archive create process into two steps

We need two types of kernel archives.

For bazel builds we need an archive of kernel headers for building
out of tree kernel modules.  Also this archive must be made public to
work with bazel.

For making VMs we need an archive of a Debian repo containing .deb
packages.  This archive should remain private as it contains our
compiled kernel artifacts and kernel source .deb packages.

This commit splits the original workflow into two archives.

Signed-off-by: Curt Brune <curt@enfabrica.net>