bestie: Add test metrics processing to BEP server
This change allows the BEP server to process TestResult
events containing test metrics presented in the
test_metrics.pb file that is incorporated in the
outputs.zip file. The test_metrics.proto defines
the protobuf message format, which reflects the
data format needed by BigQuery for Prometheus consumption.

A follow on PR is needed to continue this work, taking
the test metric data and storing it in BigQuery. For now,
the BEP server continues to display the Bazel event info
same as before, and additionally displays the relevant
metric information it finds in the TestResult events.

Tested: Started server, ran 'bazel build //bestie/...
--bes_backend=grpc://localhost:6433' and saw messages
printed on stdout; bazel exited with no error.

Jira: INFRA-326