flextape client: Move to separate binary (#396)
The original license manager client has two code paths: one that talks
to the old server, and one that talks to the new server. The old server
is no longer in use.

This change copies the code path that talks to the new server into a new
binary under //flextape/client/flextape_client. This will allow us to
delete the entire //manager subtree in a future change.

This change prepares us to add signal handling required for INFRA-418
more cleanly.

Jira: INFRA-418