bestie: Unzip test metrics directly into variable
Instead of extracting the contents of outputs.zip into an
exact replica of the directory and file structure, read
the contents of test_metrics.pb and test_summary.pb into
a byte slice variable and unmarshall the protobuf message
from there. This eliminates contention for the temporary
file name and is more efficient.

Changed the file name search through outputs.zip to use
regular expressions, accepting and processing all files
named "test_metrics*.pb" and "test_summary*.pb", with the
restriction that the * part begin with _ or -. For example,
"test_metrics_foo.pb" or "test_summary-bar.pb". This is in
anticipation that a given 'bazel test' command may need to
output multiple metric protobuf messages, one per output file.